+++ app/src/test/java/com/codesoom/assignment/task/handler/TaskHttpHandlerTest.java
@@ -22,32 +22,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class TaskHttpHandlerTest {
 
-    private static TaskRepository taskRepository = new TaskRepository();
-    TaskHttpHandler taskHttpHandler = new TaskHttpHandler();
-    ObjectMapper objectMapper = new ObjectMapper();
-
-
     @Test
-    @DisplayName("Task 객체로 변경 테스트")
-    void testToTask() throws JsonProcessingException {
-        String content = "{\"id\":\"1\", \"title\":\"작업하기\"}";
+    @DisplayName("/tasks/{id} 파싱 테스트")
+    void pasingPath() {
 
-        Task task = objectMapper.readValue(content, Task.class);
-        assertNotNull(task);
-
-        System.out.println(task);
-    }
-
-    @Test
-    @DisplayName("Json 으로 변경 테스트")
-    void testToJson() throws IOException {
-        Task task = Task.builder().id(10L).title("테스트하기").build();
-
-        OutputStream outputStream = new ByteArrayOutputStream();
-        objectMapper.writeValue(outputStream, task);
-
-        assertNotNull(outputStream);
-
-        System.out.println(outputStream.toString());
     }
 }
